refactor user class, add control login & new user
