+++ src/db/models/user.js
@@ -3,9 +3,10 @@ const Schema = mongoose.Schema
 
 const userModel = new Schema({
   name: { type: String },
-  email: { type: String },
+  email: { type: String, required: true, unique: true },
   avatar: { type: String },
-  active: { type: Boolean }
+  active: { type: Boolean },
+  password: { type: String } // TODO: add encrypter (mongoose-encrypter)
 })
 
 export default mongoose.model('user', userModel)
