+++ src/user/config.js
@@ -1,0 +1,5 @@
+const config = {
+  host: 'http://localhost:3000/api/user/'
+}
+
+export default config
